Get client ui locale on windows natively
