+++ client/ui/font_linux.go
@@ -1,0 +1,5 @@
+package main
+
+func (s *serviceClient) setDefaultFonts() {
+	//TODO: Linux Multiple Language Support
+}
